Reformatted the file structure in Obsidian
Reformatted the file structure in Obsidian
Stats and Weapons & Armor folders.
Each of those have Player and Enemy folders for their Stats or Weapons & Armor that they'll use.
The Enemy folder for both the Stats and Weapons & Armor is broken down into each sector. Each sector has a canvas to comprise everything into a mood board and have a better picture of scope. Each weapon, armor piece, and enemy (of different stat) should be a new note that is used for the canvas.